Move TentBlur1D processor to layers and replace LockTextureProxy with makeTextureImage
