+++ src/layers/processors/GLSLTentBlur1DFragmentProcessor.h
@@ -18,7 +18,7 @@
 
 #pragma once
 
-#include "gpu/processors/TentBlur1DFragmentProcessor.h"
+#include "layers/processors/TentBlur1DFragmentProcessor.h"
 
 namespace tgfx {
